Fix glitch in Cellphones; Add numerous Canadian cities and provinces

+++ catscripts/Government/Countries/Azerbaijan/Azerbaijan.sh
@@ -1,14 +1,31 @@
 #!/bin/bash
 
-egrep -i 'Azerbaijan' newpages.txt >> Azerbaijan.txt
+KEYWORDS_AZERBAIJAN="Azerbaijan"
+KEYWORDS_AZERBAIJAN_ALL="$KEYWORDS_AZERBAIJAN"
 
-AZERBAIJAN=`stat --print=%s Azerbaijan.txt`
-
-if [ $AZERBAIJAN -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="Azerbaijan.txt"
-  export CATNAME="Azerbaijan"
-  $CATEGORIZE
-fi
+  
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Azerbaijan\n"
+  fi
 
-rm Azerbaijan.txt
+  AZERBAIJAN=`egrep -i "$KEYWORDS_AZERBAIJAN" newpages.txt`
+
+  if [ "$AZERBAIJAN" != "" ];
+  then
+    printf "$AZERBAIJAN" > Azerbaijan.txt
+    export CATFILE="Azerbaijan.txt"
+    export CATNAME="Azerbaijan"
+    $CATEGORIZE
+    rm Azerbaijan.txt
+    unset AZERBAIJAN
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Azerbaijan\n"
+  fi
+
+fi
